requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.2
Django==4.1.5
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==17.4.0
iniconfig==2.0.0
packaging==23.0
paramiko==3.1.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyNaCl==1.5.0
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
sshtunnel==0.4.0
